pmoves/services/pmoves-yt/requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
curl-cffi>=0.7.0
curl-cffi==0.7.1
boto3==1.34.162
requests==2.32.4
nats-py==2.7.2
jsonschema==4.22.0
httpx==0.27.0
tenacity==8.4.2
bgutil-ytdlp-pot-provider==1.2.2
pytest-asyncio==0.23.7
